[step2] refactor: Caption 컴포넌트 style 수정
